+++ examples/aco_tsp.rs
@@ -1,8 +1,8 @@
 fn main() {
-  let (cities, cost) = ecrs::aco::generate_tsp_cost(30);
-  ecrs::aco::write_cities_csv(&cities, "cities.csv").expect("Error while writing city file");
+  let (cities, cost) = ecrs::aco::util::generate_tsp_cost(30);
+  ecrs::aco::util::write_cities_csv(&cities, "cities.csv").expect("Error while writing city file");
 
-  let heuristic = ecrs::aco::create_heuristic_from_weights(&cost);
+  let heuristic = ecrs::aco::util::create_heuristic_from_weights(&cost);
 
   let ant_s = ecrs::aco::Builder::new()
     .set_weights(cost)
