--- src/aco/ants_system.rs
@@ -1,1 +1,0 @@
-
